+++ NEWS.md
@@ -1,5 +1,9 @@
 # TidierData.jl updates
 
+## v.0.17.0 - 2025-01-27
+- add logging ability to track changes to dataframe with `TidierData_set("log", true)`
+- add docs around logging and code printing
+
 ## v0.16.5 - 2025-01-11
 - Bugfix: Corrected bug when using `Module.function()` syntax within expressions, which was previously causing errors due to the module being escaped.